refactor(feature-flag): move SchemaValidator to validator package
